RF: prepare and relocate shared helpers for reuse by the upcoming `datalad reset` command + create reset.py to host the reset-related ones.
- `_annex_reset_target` / `_reset_hard`: drop their refuse-if-dirty check and 
  relocate that guard into `update --how=reset`'s dispatch. The helpers now always 
  discard like `git reset --hard` (the git-parity `reset` wants); `update` keeps its
  stricter refuse-if-dirty contract by guarding before it delegates.
- Give `_annex_reset_target` `--force` on the corresponding-branch checkout so
  a dirty tree is discarded -- git-parity for `reset`; a no-op for `update`,
  which refuses dirty before reaching here.
- Move both helpers together with `_get_adjust_mode` into `reset.py`.
- `corresponding_hexsha`: previously introduced for testing on adjusted, now 
promote into production code into `datalad.distribution.utils`.
- `_try_command`: generic result-wrapper, move out of update.py into
  `datalad/distribution/utils.py`.
